[spirit] Split instructions into new module
