+++ spirit/src/cpu/mod.rs
@@ -6,22 +6,22 @@ use serde::Deserialize;
 use serde::Serialize;
 use tracing::trace;
 
-use crate::lookup::ArithmeticOp;
-use crate::lookup::BitOp;
-use crate::lookup::BitOpInner;
-use crate::lookup::BitShiftOp;
-use crate::lookup::Condition;
-use crate::lookup::ControlOp;
-use crate::lookup::HalfRegister;
-use crate::lookup::Instruction;
-use crate::lookup::InterruptOp;
-use crate::lookup::JumpOp;
-use crate::lookup::LoadAPointer;
-use crate::lookup::LoadOp;
-use crate::lookup::RegOrPointer;
-use crate::lookup::SomeByte;
-use crate::lookup::WideReg;
-use crate::lookup::WideRegWithoutSP;
+use crate::instruction::ArithmeticOp;
+use crate::instruction::BitOp;
+use crate::instruction::BitOpInner;
+use crate::instruction::BitShiftOp;
+use crate::instruction::Condition;
+use crate::instruction::ControlOp;
+use crate::instruction::HalfRegister;
+use crate::instruction::Instruction;
+use crate::instruction::InterruptOp;
+use crate::instruction::JumpOp;
+use crate::instruction::LoadAPointer;
+use crate::instruction::LoadOp;
+use crate::instruction::RegOrPointer;
+use crate::instruction::SomeByte;
+use crate::instruction::WideReg;
+use crate::instruction::WideRegWithoutSP;
 use crate::mem::MemoryLikeExt;
 use crate::utils::Wrapping;
