+++ spirit/src/lookup.rs
@@ -1,12 +1,10 @@
 use array_concat::concat_arrays;
 
-use crate::cpu::Cpu;
+use crate::instruction::*;
+// #[cfg(test)]
 use crate::mem::MemoryLike;
 #[cfg(not(test))]
 use crate::mem::MemoryMap;
-
-use derive_more::From;
-use derive_more::IsVariant;
 
 #[cfg(test)]
 #[track_caller]
@@ -37,640 +35,22 @@ fn parse_prefixed_instruction(mem: &MemoryMap, pc: u16) -> Instruction {
 
 #[cfg(not(test))]
 type OpArray<const N: usize> = [fn(&MemoryMap, u16) -> Instruction; N];
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum Instruction {
-    #[display("{_0}")]
-    Load(LoadOp),
-    #[display("{_0}")]
-    BitShift(BitShiftOp),
-    #[display("{_0}")]
-    ControlOp(ControlOp),
-    #[display("{_0}")]
-    Bit(BitOp),
-    #[display("{_0}")]
-    Jump(JumpOp),
-    #[display("{_0}")]
-    Arithmetic(ArithmeticOp),
-    #[display("Interrupt {_0}")]
-    Interrupt(InterruptOp),
-    #[display("DAA")]
-    Daa,
-    /// Set Carry.
-    #[display("SCF")]
-    Scf,
-    /// ComPLement accumulator.
-    #[display("CPL")]
-    Cpl,
-    /// CompLement carry flag.
-    #[display("CCF")]
-    Ccf,
-    /// Disable interupts
-    #[display("DI")]
-    Di,
-    /// Enable interupts
-    #[display("EI")]
-    Ei,
-    /// This is not a real instruction, but its used to communicate that the VRAM DMA is
-    /// transferring data. One of these operations represents 16 bytes being transferred.
-    Transfer,
+#[cfg(test)]
+#[track_caller]
+fn invalid_op_code(mem: &dyn MemoryLike, pc: u16) -> Instruction {
+    unreachable!(
+        "Op code '0x{:0>2X}' @ 0x{pc:0>4X} does not correspond to any valid operation",
+        mem.read_byte(pc)
+    )
 }
 
-impl Instruction {
-    /// Returns the number of ticks to will take to complete this instruction.
-    /// Takes a reference to the CPU in order to determine if this instruction will pass any
-    /// conditions.
-    pub fn length(&self, cpu: &Cpu) -> u8 {
-        match self {
-            Instruction::Load(op) => op.length(),
-            Instruction::BitShift(op) => op.length(),
-            Instruction::ControlOp(op) => op.length(),
-            Instruction::Bit(op) => op.length(),
-            Instruction::Jump(op) => op.length(cpu),
-            Instruction::Arithmetic(op) => op.length(),
-            Instruction::Interrupt(_) => 20, // Interrupts are, basically, calls but last 5 cycles
-            Instruction::Daa => 4,
-            Instruction::Scf => 4,
-            Instruction::Cpl => 4,
-            Instruction::Ccf => 4,
-            Instruction::Di => 4,
-            Instruction::Ei => 4,
-            // It takes 24 ticks to transfer 16 bytes.
-            Instruction::Transfer => 24,
-        }
-    }
-
-    /// Returns the size of the bytes to took to construct this instruction
-    pub const fn size(&self) -> u8 {
-        match self {
-            Instruction::Load(op) => op.size(),
-            Instruction::BitShift(op) => op.size(),
-            Instruction::ControlOp(op) => op.size(),
-            Instruction::Bit(op) => op.size(),
-            Instruction::Jump(op) => op.size(),
-            Instruction::Arithmetic(op) => op.size(),
-            Instruction::Interrupt(_) => 0, // Interrupts don't move the PC
-            Instruction::Daa => 1,
-            Instruction::Scf => 1,
-            Instruction::Cpl => 1,
-            Instruction::Ccf => 1,
-            Instruction::Di => 1,
-            Instruction::Ei => 1,
-            // This is a fake instruction and it should not move the PC
-            Instruction::Transfer => 0,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-#[repr(u16)]
-pub enum InterruptOp {
-    #[display("VBlank")]
-    VBlank = 0x0040,
-    #[display("LCD")]
-    LCD = 0x0048,
-    #[display("Timer")]
-    Timer = 0x0050,
-    #[display("Serial")]
-    Serial = 0x0058,
-    #[display("Joypad")]
-    Joypad = 0x0060,
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::From, derive_more::Display)]
-#[display("{_variant}")]
-pub enum SomeByte {
-    #[display("{_0}")]
-    Referenced(RegOrPointer),
-    #[display("0x{_0:0>2X}")]
-    Direct(u8),
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum ArithmeticOp {
-    #[display("ADD {_0}")]
-    Add(SomeByte),
-    #[display("ADD HL, {_0}")]
-    Add16(WideReg),
-    #[display("ADD SP, {_0}")]
-    AddSP(i8),
-    #[display("ADC {_0}")]
-    Adc(SomeByte),
-    #[display("SUB {_0}")]
-    Sub(SomeByte),
-    #[display("SBC {_0}")]
-    Sbc(SomeByte),
-    #[display("AND {_0}")]
-    And(SomeByte),
-    #[display("XOR {_0}")]
-    Xor(SomeByte),
-    #[display("OR {_0}")]
-    Or(SomeByte),
-    #[display("CP {_0}")]
-    Cp(SomeByte),
-    #[display("INC {_0}")]
-    Inc(RegOrPointer),
-    #[display("INC {_0}")]
-    Inc16(WideReg),
-    #[display("DEC {_0}")]
-    Dec(RegOrPointer),
-    #[display("DEC {_0}")]
-    Dec16(WideReg),
-}
-
-impl ArithmeticOp {
-    /// Returns the number of ticks to will take to complete this instruction.
-    pub fn length(&self) -> u8 {
-        match self {
-            ArithmeticOp::Add16(_) => 8,
-            ArithmeticOp::Add(SomeByte::Referenced(RegOrPointer::Reg(_))) => 4,
-            ArithmeticOp::Add(_) => 8,
-            ArithmeticOp::Adc(SomeByte::Referenced(RegOrPointer::Reg(_))) => 4,
-            ArithmeticOp::Adc(_) => 8,
-            ArithmeticOp::Sub(SomeByte::Referenced(RegOrPointer::Reg(_))) => 4,
-            ArithmeticOp::Sub(_) => 8,
-            ArithmeticOp::Sbc(SomeByte::Referenced(RegOrPointer::Reg(_))) => 4,
-            ArithmeticOp::Sbc(_) => 8,
-            ArithmeticOp::And(SomeByte::Referenced(RegOrPointer::Reg(_))) => 4,
-            ArithmeticOp::And(_) => 8,
-            ArithmeticOp::Xor(SomeByte::Referenced(RegOrPointer::Reg(_))) => 4,
-            ArithmeticOp::Xor(_) => 8,
-            ArithmeticOp::Or(SomeByte::Referenced(RegOrPointer::Reg(_))) => 4,
-            ArithmeticOp::Or(_) => 8,
-            ArithmeticOp::Cp(SomeByte::Referenced(RegOrPointer::Reg(_))) => 4,
-            ArithmeticOp::Cp(_) => 8,
-            ArithmeticOp::Inc(RegOrPointer::Pointer) => 12,
-            ArithmeticOp::Inc(RegOrPointer::Reg(_)) => 4,
-            ArithmeticOp::Dec(RegOrPointer::Pointer) => 12,
-            ArithmeticOp::Dec(RegOrPointer::Reg(_)) => 4,
-            ArithmeticOp::Inc16(_) => 8,
-            ArithmeticOp::Dec16(_) => 8,
-            ArithmeticOp::AddSP(_) => 16,
-        }
-    }
-
-    /// Returns the size of the bytes to took to construct this instruction
-    pub const fn size(&self) -> u8 {
-        match self {
-            ArithmeticOp::Add16(_) => 1,
-            ArithmeticOp::Add(SomeByte::Direct(_)) => 2,
-            ArithmeticOp::Add(_) => 1,
-            ArithmeticOp::Adc(SomeByte::Direct(_)) => 2,
-            ArithmeticOp::Adc(_) => 1,
-            ArithmeticOp::Sub(SomeByte::Direct(_)) => 2,
-            ArithmeticOp::Sub(_) => 1,
-            ArithmeticOp::Sbc(SomeByte::Direct(_)) => 2,
-            ArithmeticOp::Sbc(_) => 1,
-            ArithmeticOp::And(SomeByte::Direct(_)) => 2,
-            ArithmeticOp::And(_) => 1,
-            ArithmeticOp::Xor(SomeByte::Direct(_)) => 2,
-            ArithmeticOp::Xor(_) => 1,
-            ArithmeticOp::Or(SomeByte::Direct(_)) => 2,
-            ArithmeticOp::Or(_) => 1,
-            ArithmeticOp::Cp(SomeByte::Direct(_)) => 2,
-            ArithmeticOp::Cp(_) => 1,
-            ArithmeticOp::Inc(_) => 1,
-            ArithmeticOp::Dec(_) => 1,
-            ArithmeticOp::Inc16(_) => 1,
-            ArithmeticOp::Dec16(_) => 1,
-            ArithmeticOp::AddSP(_) => 2,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum JumpOp {
-    /// Op Codes: 0x20, 0x30, 0x28, 0x38
-    #[display("JR {_0}, 0x{_1:0>2X}")]
-    ConditionalRelative(Condition, i8),
-    /// Op Code: 0x18
-    #[display("JR 0x{_0:0>2X}")]
-    Relative(i8),
-    /// Op Codes: 0xC2, 0xD2, 0xCA, 0xDA
-    #[display("JP {_0}, 0x{_1:0>4X}")]
-    ConditionalAbsolute(Condition, u16),
-    /// Op Code: 0xC3
-    #[display("JP 0x{_0:0>2X}")]
-    Absolute(u16),
-    /// Op Code: 0xE9
-    #[display("JP HL")]
-    JumpToHL,
-    /// Op Code: 0xCD
-    #[display("CALL 0x{_0:0>4X}")]
-    Call(u16),
-    /// Op Codes: 0xC4, 0xD4, 0xCC, 0xDC
-    #[display("CALL {_0}, 0x{_1:0>4X}")]
-    ConditionalCall(Condition, u16),
-    /// Op Code: 0xC9
-    #[display("RET")]
-    Return,
-    /// Op Codes: 0xC0, 0xD0, 0xC8, 0xD8
-    #[display("RET {_0}")]
-    ConditionalReturn(Condition),
-    /// Op Code: 0xD9
-    /// Return from the subroutine and enable intrupts
-    #[display("RETI")]
-    ReturnAndEnable,
-    /// Op Code: 0xC7
-    #[display("RST 0x00")]
-    RST00,
-    /// Op Code: 0xCF
-    #[display("RST 0x08")]
-    RST08,
-    /// Op Code: 0xD7
-    #[display("RST 0x10")]
-    RST10,
-    /// Op Code: 0xDF
-    #[display("RST 0x18")]
-    RST18,
-    /// Op Code: 0xE7
-    #[display("RST 0x20")]
-    RST20,
-    /// Op Code: 0xEF
-    #[display("RST 0x28")]
-    RST28,
-    /// Op Code: 0xF7
-    #[display("RST 0x30")]
-    RST30,
-    /// Op Code: 0xFF
-    #[display("RST 0x38")]
-    RST38,
-}
-
-impl JumpOp {
-    /// Returns the number of ticks to will take to complete this instruction.
-    pub fn length(&self, cpu: &Cpu) -> u8 {
-        match self {
-            JumpOp::ConditionalRelative(cond, _) => 8 + (4 * cond.passed(cpu) as u8),
-            JumpOp::Relative(_) => 12,
-            JumpOp::ConditionalAbsolute(cond, _) => 12 + (4 * cond.passed(cpu) as u8),
-            JumpOp::Absolute(_) => 16,
-            JumpOp::JumpToHL => 4,
-            JumpOp::Call(_) => 24,
-            JumpOp::ConditionalCall(cond, _) => 12 + (12 * cond.passed(cpu) as u8),
-            JumpOp::Return => 16,
-            JumpOp::ConditionalReturn(cond) => 8 + (12 * cond.passed(cpu) as u8),
-            JumpOp::ReturnAndEnable => 16,
-            JumpOp::RST00 => 16,
-            JumpOp::RST08 => 16,
-            JumpOp::RST10 => 16,
-            JumpOp::RST18 => 16,
-            JumpOp::RST20 => 16,
-            JumpOp::RST28 => 16,
-            JumpOp::RST30 => 16,
-            JumpOp::RST38 => 16,
-        }
-    }
-
-    /// Returns the size of the bytes to took to construct this instruction
-    pub const fn size(&self) -> u8 {
-        match self {
-            JumpOp::ConditionalRelative(_, _) => 2,
-            JumpOp::Relative(_) => 2,
-            JumpOp::ConditionalAbsolute(_, _) => 3,
-            JumpOp::Absolute(_) => 3,
-            JumpOp::JumpToHL => 1,
-            JumpOp::Call(_) => 3,
-            JumpOp::ConditionalCall(_, _) => 3,
-            JumpOp::Return => 1,
-            JumpOp::ConditionalReturn(_) => 1,
-            JumpOp::ReturnAndEnable => 1,
-            JumpOp::RST00 => 1,
-            JumpOp::RST08 => 1,
-            JumpOp::RST10 => 1,
-            JumpOp::RST18 => 1,
-            JumpOp::RST20 => 1,
-            JumpOp::RST28 => 1,
-            JumpOp::RST30 => 1,
-            JumpOp::RST38 => 1,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum ControlOp {
-    #[display("HALT")]
-    Halt,
-    #[display("NOOP")]
-    Noop,
-    #[display("Stop 0x{_0:0>2X}")]
-    Stop(u8),
-}
-
-impl ControlOp {
-    /// Returns the number of ticks to will take to complete this instruction.
-    pub fn length(&self) -> u8 {
-        match self {
-            ControlOp::Noop => 4,
-            ControlOp::Stop(_) => 4,
-            ControlOp::Halt => 4,
-        }
-    }
-
-    /// Returns the size of the bytes to took to construct this instruction
-    pub const fn size(&self) -> u8 {
-        match self {
-            ControlOp::Noop => 1,
-            ControlOp::Stop(_) => 2,
-            ControlOp::Halt => 1,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum WideReg {
-    #[display("BC")]
-    BC,
-    #[display("DE")]
-    DE,
-    #[display("HL")]
-    HL,
-    #[display("SP")]
-    SP,
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum WideRegWithoutSP {
-    #[display("BC")]
-    BC,
-    #[display("DE")]
-    DE,
-    #[display("HL")]
-    HL,
-    #[display("AF")]
-    AF,
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum LoadOp {
-    /// Used for opcodes in 0x40..0x80
-    #[display("LD {src} -> {dest}")]
-    Basic {
-        dest: RegOrPointer,
-        src: RegOrPointer,
-    },
-    /// Used for opcodes 0xX1
-    #[display("LD 0x{_1:0>4X} -> {_0}")]
-    Direct16(WideReg, u16),
-    /// Used for opcodes 0x_6 and 0x_E
-    #[display("LD 0x{_1:0>2X} -> {_0}")]
-    Direct(RegOrPointer, u8),
-    /// Used for opcodes 0x_A
-    #[display("LD {_0} -> A")]
-    LoadIntoA(LoadAPointer),
-    /// Used for opcodes 0x_2
-    #[display("LD A -> {_0}")]
-    StoreFromA(LoadAPointer),
-    /// Opcode: 0x08
-    /// Store SP & $FF at address n16 and SP >> 8 at address n16 + 1.
-    #[display("LD SP -> 0x{_0:0>4X}")]
-    StoreSP(u16),
-    /// Opcode: 0xF9
-    #[display("LD SP -> HL")]
-    HLIntoSP,
-    /// Opcode: 0xF8
-    /// Add the signed value e8 to SP and store the result in HL.
-    #[display("LD SP + {_0} -> HL")]
-    SPIntoHL(i8),
-    /// Used for opcodes 0x_1
-    #[display("POP {_0}")]
-    Pop(WideRegWithoutSP),
-    /// Used for opcodes 0x_5
-    #[display("PUSH {_0}")]
-    Push(WideRegWithoutSP),
-    /// Used for opcode 0xE0
-    #[display("LDH A -> 0xFF{_0:0>2X}")]
-    LoadHigh(u8),
-    /// Used for opcode 0xF0
-    #[display("LDH 0xFF{_0:0>2X} -> A")]
-    StoreHigh(u8),
-    /// Used for opcode 0xE2
-    #[display("LDHCA")]
-    Ldhca,
-    /// Used for opcode 0xF2
-    #[display("LDHAC")]
-    Ldhac,
-    // TODO: These are backwards. For `LoadA`, the ptr is the dest and A is the value. The reverse
-    // is true for `StoreA`.
-    /// Used for opcode 0xEA
-    #[display("LD 0x{ptr:0>4X} -> A")]
-    LoadA { ptr: u16 },
-    /// Used for opcode 0xFA
-    #[display("LD A -> 0x{ptr:0>4X}")]
-    StoreA { ptr: u16 },
-}
-
-impl LoadOp {
-    /// Returns the number of ticks to will take to complete this instruction.
-    pub fn length(&self) -> u8 {
-        match self {
-            LoadOp::Basic {
-                dest: RegOrPointer::Pointer,
-                src: RegOrPointer::Pointer,
-            } => 4,
-            LoadOp::Basic {
-                dest: RegOrPointer::Pointer,
-                ..
-            }
-            | LoadOp::Basic {
-                src: RegOrPointer::Pointer,
-                ..
-            } => 8,
-            LoadOp::Basic { .. } => 4,
-            LoadOp::Direct16(_, _) => 12,
-            LoadOp::Direct(RegOrPointer::Pointer, _) => 12,
-            LoadOp::Direct(_, _) => 8,
-            LoadOp::LoadIntoA(_) => 8,
-            LoadOp::StoreFromA(_) => 8,
-            LoadOp::StoreSP(_) => 20,
-            LoadOp::HLIntoSP => 8,
-            LoadOp::SPIntoHL(_) => 12,
-            LoadOp::Pop(_) => 12,
-            LoadOp::Push(_) => 16,
-            LoadOp::LoadHigh(_) => 12,
-            LoadOp::StoreHigh(_) => 12,
-            LoadOp::Ldhca => 8,
-            LoadOp::Ldhac => 8,
-            LoadOp::LoadA { .. } => 16,
-            LoadOp::StoreA { .. } => 16,
-        }
-    }
-
-    /// Returns the size of the bytes to took to construct this instruction
-    pub const fn size(&self) -> u8 {
-        match self {
-            LoadOp::Basic { .. } => 1,
-            LoadOp::Direct16(_, _) => 3,
-            LoadOp::Direct(_, _) => 2,
-            LoadOp::LoadIntoA(_) => 1,
-            LoadOp::StoreFromA(_) => 1,
-            LoadOp::StoreSP(_) => 3,
-            LoadOp::HLIntoSP => 1,
-            LoadOp::SPIntoHL(_) => 2,
-            LoadOp::Pop(_) => 1,
-            LoadOp::Push(_) => 1,
-            LoadOp::LoadHigh(_) => 2,
-            LoadOp::StoreHigh(_) => 2,
-            LoadOp::Ldhca => 1,
-            LoadOp::Ldhac => 1,
-            LoadOp::LoadA { .. } => 3,
-            LoadOp::StoreA { .. } => 3,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum Condition {
-    #[display("Z")]
-    Zero,
-    #[display("NZ")]
-    NotZero,
-    #[display("C")]
-    Carry,
-    #[display("NC")]
-    NotCarry,
-}
-
-impl Condition {
-    pub fn passed(&self, cpu: &Cpu) -> bool {
-        match self {
-            Condition::Zero => cpu.zero_flag(),
-            Condition::NotZero => !cpu.zero_flag(),
-            Condition::Carry => cpu.carry_flag(),
-            Condition::NotCarry => !cpu.carry_flag(),
-        }
-    }
-}
-
-/// There are special operations for loading into the A register, so it is easier to have a special
-/// enum for the unique types of pointers they use.
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum LoadAPointer {
-    /// Use the BC register
-    #[display("BC")]
-    BC,
-    /// Use the DE register
-    #[display("DE")]
-    DE,
-    /// Use the HL register and increment after performing the operation
-    #[display("HL+")]
-    Hli,
-    /// Use the HL register and decrement after performing the operation
-    #[display("HL-")]
-    Hld,
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum BitShiftOp {
-    #[display("RLC {_0}")]
-    Rlc(RegOrPointer),
-    #[display("RLCA")]
-    Rlca,
-    #[display("RRC {_0}")]
-    Rrc(RegOrPointer),
-    #[display("RRCA")]
-    Rrca,
-    #[display("RL {_0}")]
-    Rl(RegOrPointer),
-    #[display("RLA")]
-    Rla,
-    #[display("RR {_0}")]
-    Rr(RegOrPointer),
-    #[display("RRA")]
-    Rra,
-    #[display("SLA {_0}")]
-    Sla(RegOrPointer),
-    #[display("SRA {_0}")]
-    Sra(RegOrPointer),
-    #[display("SWAP {_0}")]
-    Swap(RegOrPointer),
-    #[display("SRL {_0}")]
-    Srl(RegOrPointer),
-}
-
-impl BitShiftOp {
-    /// Returns the number of ticks to will take to complete this instruction.
-    pub fn length(&self) -> u8 {
-        match self {
-            BitShiftOp::Rlca => 4,
-            BitShiftOp::Rlc(RegOrPointer::Pointer) => 16,
-            BitShiftOp::Rlc(_) => 8,
-            BitShiftOp::Rrca => 4,
-            BitShiftOp::Rrc(RegOrPointer::Pointer) => 16,
-            BitShiftOp::Rrc(_) => 8,
-            BitShiftOp::Rl(RegOrPointer::Pointer) => 16,
-            BitShiftOp::Rl(_) => 8,
-            BitShiftOp::Rla => 4,
-            BitShiftOp::Rr(RegOrPointer::Pointer) => 16,
-            BitShiftOp::Rr(_) => 8,
-            BitShiftOp::Rra => 4,
-            BitShiftOp::Sla(RegOrPointer::Pointer) => 16,
-            BitShiftOp::Sla(_) => 8,
-            BitShiftOp::Sra(RegOrPointer::Pointer) => 16,
-            BitShiftOp::Sra(_) => 8,
-            BitShiftOp::Swap(RegOrPointer::Pointer) => 16,
-            BitShiftOp::Swap(_) => 8,
-            BitShiftOp::Srl(RegOrPointer::Pointer) => 16,
-            BitShiftOp::Srl(_) => 8,
-        }
-    }
-
-    /// Returns the size of the bytes to took to construct this instruction
-    pub const fn size(&self) -> u8 {
-        match self {
-            BitShiftOp::Rlca => 1,
-            BitShiftOp::Rlc(_) => 2,
-            BitShiftOp::Rrca => 1,
-            BitShiftOp::Rrc(_) => 2,
-            BitShiftOp::Rla => 1,
-            BitShiftOp::Rl(_) => 2,
-            BitShiftOp::Rra => 1,
-            BitShiftOp::Rr(_) => 2,
-            BitShiftOp::Sla(_) => 2,
-            BitShiftOp::Sra(_) => 2,
-            BitShiftOp::Swap(_) => 2,
-            BitShiftOp::Srl(_) => 2,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum HalfRegister {
-    #[display("A")]
-    A,
-    #[display("B")]
-    B,
-    #[display("C")]
-    C,
-    #[display("D")]
-    D,
-    #[display("E")]
-    E,
-    #[display("H")]
-    H,
-    #[display("L")]
-    L,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, From, IsVariant, derive_more::Display)]
-#[display("{_variant}")]
-pub enum RegOrPointer {
-    #[display("{_0}")]
-    Reg(HalfRegister),
-    #[display("Pointer")]
-    Pointer,
+#[cfg(not(test))]
+#[track_caller]
+fn invalid_op_code(mem: &MemoryMap, pc: u16) -> Instruction {
+    unreachable!(
+        "Op code '0x{:0>2X}' @ 0x{pc:0>4X} does not correspond to any valid operation",
+        mem.read_byte(pc)
+    )
 }
 
 enum InnerRegOrPointer {
@@ -697,55 +77,6 @@ impl InnerRegOrPointer {
             InnerRegOrPointer::Pointer => RegOrPointer::Pointer,
         }
     }
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{op} {bit} {reg}")]
-pub struct BitOp {
-    pub bit: u8,
-    pub reg: RegOrPointer,
-    pub op: BitOpInner,
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, derive_more::Display)]
-#[display("{_variant}")]
-pub enum BitOpInner {
-    #[display("BIT")]
-    Bit,
-    #[display("RES")]
-    Res,
-    #[display("SET")]
-    Set,
-}
-
-impl BitOp {
-    /// Returns the number of ticks to will take to complete this instruction.
-    pub fn length(&self) -> u8 {
-        8 + 8 * (self.reg.is_pointer() as u8)
-    }
-
-    /// Returns the size of the bytes to took to construct this instruction
-    pub const fn size(&self) -> u8 {
-        2
-    }
-}
-
-#[cfg(test)]
-#[track_caller]
-fn invalid_op_code(mem: &dyn MemoryLike, pc: u16) -> Instruction {
-    unreachable!(
-        "Op code '0x{:0>2X}' @ 0x{pc:0>4X} does not correspond to any valid operation",
-        mem.read_byte(pc)
-    )
-}
-
-#[cfg(not(test))]
-#[track_caller]
-fn invalid_op_code(mem: &MemoryMap, pc: u16) -> Instruction {
-    unreachable!(
-        "Op code '0x{:0>2X}' @ 0x{pc:0>4X} does not correspond to any valid operation",
-        mem.read_byte(pc)
-    )
 }
 
 macro_rules! define_op {
